test: move Lua script Redis tests out of the mock-only test_litellm tree
Greptile finding: tests/test_litellm/readme.md documents that directory as
mocked-tests-only, but the six new Lua script tests contacted a real
redis-server subprocess. Moved them verbatim to
tests/local_testing/test_tag_rate_limits_shared_redis.py, which already
hosts other real-Redis tests (test_redis_batch_optimizations.py). No
behavior change; tests/test_litellm/proxy/hooks/test_tag_rate_limits_shared.py
is mock-only again.
